+++ extensions/iceworks-doctor/web/src/pages/Dashboard/index.tsx
@@ -6,11 +6,15 @@ import ScanCard from './components/ScanCard';
 const { Cell } = ResponsiveGrid;
 
 const Dashboard = () => {
+  // @ts-ignore
+  const autoScan = !!window.AUTO_SCAN;
   return (
     <ResponsiveGrid gap={10}>
-      <Cell colSpan={12}>
-        <InfoCard />
-      </Cell>
+      {autoScan ? null : (
+        <Cell colSpan={12}>
+          <InfoCard />
+        </Cell>
+      )}
       <Cell colSpan={12}>
         <ScanCard />
       </Cell>